Refactor out training-test-util
