+++ training-test-util/training-test-util.gradle.kts
@@ -1,0 +1,9 @@
+description = "Utilities for testing code that trains models."
+
+dependencies {
+    api(project(":test-util"))
+    api(project(":dsl"))
+    api(project(":tf-layer-loader"))
+
+    implementation(project(":logging"))
+}
