+++ training/src/test/kotlin/edu/wpi/axon/training/TrainSequentialModelScriptGeneratorIntegrationTest.kt
@@ -8,6 +8,8 @@ import edu.wpi.axon.tfdata.Dataset
 import edu.wpi.axon.tfdata.Model
 import edu.wpi.axon.tfdata.loss.Loss
 import edu.wpi.axon.tfdata.optimizer.Optimizer
+import edu.wpi.axon.training.testutil.loadModel
+import edu.wpi.axon.training.testutil.testTrainingScript
 import io.kotlintest.assertions.arrow.validation.shouldBeInvalid
 import io.kotlintest.assertions.arrow.validation.shouldBeValid
 import io.kotlintest.matchers.types.shouldBeInstanceOf
