+++ ui-vaadin/src/main/kotlin/edu/wpi/axon/ui/model/Job.kt
@@ -5,5 +5,3 @@ import edu.wpi.axon.tfdata.Dataset
 enum class JobState { NOT_STARTED, IN_PROGRESS, COMPLETED }
 
 data class Job(var name: String, var state: JobState, var dataset: Dataset)
-
-
