+++ ui-vaadin/src/main/kotlin/edu/wpi/axon/ui/view/ModelView.kt
@@ -28,4 +28,3 @@ class ModelView : KComposite() {
 //        }
 //    }
 }
-
