+++ ui-vaadin/src/test/kotlin/edu/wpi/axon/ui/JobRunnerTest.kt
@@ -3,18 +3,13 @@ package edu.wpi.axon.ui
 import edu.wpi.axon.dbdata.Job
 import edu.wpi.axon.dbdata.TrainingScriptProgress
 import edu.wpi.axon.tfdata.Dataset
-import edu.wpi.axon.tfdata.Model
 import edu.wpi.axon.tfdata.loss.Loss
 import edu.wpi.axon.tfdata.optimizer.Optimizer
-import edu.wpi.axon.tflayerloader.DefaultLayersToGraph
-import edu.wpi.axon.tflayerloader.LoadLayersFromHDF5
-import io.kotlintest.assertions.arrow.either.shouldBeRight
+import edu.wpi.axon.training.testutil.loadModel
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import software.amazon.awssdk.regions.Region
 import software.amazon.awssdk.services.ec2.model.InstanceType
-import java.io.File
-import java.nio.file.Paths
 
 internal class JobRunnerTest {
 
@@ -44,14 +39,4 @@ internal class JobRunnerTest {
 
         jobRunner.startJob(job)
     }
-}
-
-// TODO: Put this into a test util project so it's not copy-pasted
-internal fun loadModel(modelName: String, stub: () -> Unit = {}): Pair<Model, String> {
-    val localModelPath = Paths.get(stub::class.java.getResource(modelName).toURI()).toString()
-    val layers = LoadLayersFromHDF5(DefaultLayersToGraph())
-        .load(File(localModelPath))
-    val model = layers.attempt().unsafeRunSync()
-    model.shouldBeRight()
-    return model.b as Model to localModelPath
 }
